Added Get-OktaApp + fixes after preliminary tests
