+++ Okta.Core.Automation/Okta.Core.Automation.csproj
@@ -29,7 +29,7 @@
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
+    <OutputPath>bin\</OutputPath>
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
@@ -87,10 +87,9 @@
     <Compile Include="OktaCmdlet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Users\EnrollOktaUserFactor.cs" />
-    <Compile Include="Users\GetOktaAppUsers.cs" />
     <Compile Include="Users\GetOktaAppUser.cs" />
+    <Compile Include="Users\GetOktaApp.cs" />
     <Compile Include="Users\GetOktaUserFactor.cs" />
-    <Compile Include="Users\GetOktaUserFactors.cs" />
     <Compile Include="Users\NewOktaUser.cs" />
     <Compile Include="Users\SetOktaUser.cs" />
     <Compile Include="Users\DisableOktaUser.cs" />
