+++ Okta.Core.Automation/chocolatey/okta.core.automation.nuspec
@@ -3,7 +3,7 @@
   <metadata>
     <title>Okta PowerShell Module</title>
     <id>okta.core.automation</id>
-    <version>0.1.5.6</version>
+    <version>0.2.0.0</version>
     <authors>okta</authors>
     <owners>okta</owners>
     <projectUrl>http://developer.okta.com/docs/sdk/core/api.html</projectUrl>
